r2dbc save list<Company>

+++ src/main/java/com/example/maxhodik/stock/analyze/repository/CompanyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CompanyRepository extends R2dbcRepository<Company, Integer> {
+public interface CompanyRepository extends R2dbcRepository<Company, Long> {
 //public interface CompanyRepository extends JpaRepository<Company, Long> {
 
 }
